Breakers & Bugfixes
NEW - Lazer Breaker
Gathered from Timothy, the Lazer Breaker is a addon to the ship that deals rapid 25.0 damage with a tickrate of 0.01 seconds; draining -5.0 charge every tick. Press C when at 75% charge to use in-game.

Fixed a bug where when locking in repeatedly on a enemy, the hitbox dealing the damage would multiply endlessly.
Fixed a bug where the players hit audio sometimes wouldn't play.